requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart
numpy
opencv-python-headless
gunicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]